requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-v5==1.0.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==2.1.1
pkg-resources==0.0.0
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
